+++ script.module.web-pdb/addon.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8" standalone="yes"?>
 <addon id="script.module.web-pdb"
        name="Web-PDB"
-       version="1.6.2"
+       version="1.6.3"
        provider-name="Roman V.M.">
   <requires>
     <import addon="xbmc.python" version="3.0.0"/>
@@ -22,7 +22,6 @@
       <icon>icon.png</icon>
       <screenshot>resources/screenshot.jpg</screenshot>
     </assets>
-    <news>- 1.6.2: Various internal changes.
-- 1.6.0: Dropped Python 2 compatibility.</news>
+    <news>- 1.6.3: Sync changes with the upstream library version.</news>
   </extension>
 </addon>
