+++ script.module.web-pdb/libs/web_pdb/static/styles.min.css.sha256
@@ -1,1 +1,1 @@
-c69f3afa25647e7eeb45402621a59f83d3007005fb4c54f27a19f58e4358058a  styles.min.css
+71ffc923fa5b013c1450597ae0a3443f79c26fc38f21ed7aab7b287ec883e5dd *styles.min.css
