+++ script.module.web-pdb/libs/web_pdb/web_console.py
@@ -26,13 +26,14 @@ File-like web-based input/output console
 
 import queue
 import weakref
-from threading import Thread, Event, RLock
+from threading import Thread, Event
 
 import xbmc
 from xbmcaddon import Addon
 from xbmcgui import DialogProgressBG
 
 from .asyncore_wsgi import make_server, AsyncWebSocketHandler
+from .buffer import ThreadSafeBuffer
 from .logging import getLogger
 from .wsgi_app import app
 
@@ -41,35 +42,6 @@ __all__ = ['WebConsole']
 kodi_monitor = xbmc.Monitor()
 addon = Addon('script.module.web-pdb')
 logger = getLogger(__name__)
-
-
-class ThreadSafeBuffer(object):
-    """
-    A buffer for data exchange between threads
-    """
-    def __init__(self, contents=None):
-        self._lock = RLock()
-        self._contents = contents
-        self._is_dirty = contents is not None
-
-    @property
-    def is_dirty(self):
-        """Indicates whether a buffer contains unread data"""
-        with self._lock:
-            return self._is_dirty
-
-    @property
-    def contents(self):
-        """Get or set buffer contents"""
-        with self._lock:
-            self._is_dirty = False
-            return self._contents
-
-    @contents.setter
-    def contents(self, value):
-        with self._lock:
-            self._contents = value
-            self._is_dirty = True
 
 
 class WebConsoleSocket(AsyncWebSocketHandler):
@@ -103,7 +75,7 @@ class WebConsole(object):
     def __init__(self, host, port, debugger):
         self._debugger = weakref.proxy(debugger)
         self._console_history = ThreadSafeBuffer('')
-        self._frame_data = ThreadSafeBuffer()
+        self._frame_data = None
         self._stop_all = Event()
         self._server_thread = Thread(target=self._run_server, args=(host, port))
         self._server_thread.daemon = True
@@ -126,7 +98,7 @@ class WebConsole(object):
         return self._stop_all.is_set()
 
     def _run_server(self, host, port):
-        app.frame_data = self._frame_data
+        self._frame_data = app.frame_data
         httpd = make_server(host, port, app, ws_handler_class=WebConsoleSocket)
         logger.info(f'Web-PDB: starting web-server on {httpd.server_name}:{port}...')
         dialog = DialogProgressBG()
