+++ script.module.web-pdb/libs/web_pdb/wsgi_app.py
@@ -32,6 +32,8 @@ from io import BytesIO
 
 import bottle
 
+from .buffer import ThreadSafeBuffer
+
 __all__ = ['app']
 
 # bottle.debug(True)
@@ -65,7 +67,7 @@ def compress(func):
 class WebConsoleApp(bottle.Bottle):
     def __init__(self):
         super(WebConsoleApp, self).__init__()
-        self.frame_data = None
+        self.frame_data = ThreadSafeBuffer()
 
 
 app = WebConsoleApp()
